Vendor Repo bring-up [Read Comments]
This is my first ever Vendor repo bring up, so I wasnt sure how to do it properly.
I used this: https://github.com/JackpotClavin/Android-Blob-Utility/blob/master/emulator_systems/sdk_22.txt
I also referenced LineageOS external_* repos.
With the above 2 I had removed over half the librarys becuse they where stock.
Still have a big pile so I then removed files that I thought werent needed, the list below is the files removed by myself (includes a few from the LineageOS repo list):
libasf_fileshareserver.so
libasf_fileshare.so
libknox_encryption.so
libknox_km.so
libknox_legacy_ode_km.so
libknoxvpnfdsender.so
libm.so
libmtp.so
libremotedesktop_client.so
libremotedesktopservice.so
libsamsungearcare.so
libsamsungeffect.so
libSamsungPkcs11Wrapper.so
libSamsungPostProcess.so
libsamsungpowersound.so
libSamsungPreProcess.so
libsamsungRecord.so
lib_SamsungRec_V04013.so
lib_Samsung_Resampler.so
libsamsungSoundbooster.so
lib_SoundAlive_play_ver125e.so
lib_soundaliveresampler.so
libsoundalive.so
lib_SoundAlive_SRC192_ver205.so
lib_SoundAlive_ver118t.so
libSPenBase.so
libSPenBeautify.so
libSPenBrush.so
libSPenChineseBrush.so
libSPenEngine.so
libSPenFountainPen.so
libSPenHSV.so
libSPenInit.so
libSPenInkPen.so
libSPenMagicPen.so
libSPenMarker.so
libSPenModel.so
libSPenMontblancCalligraphyPen.so
libSPenMontblancFountainPen.so
libSPenNativePen.so
libSPenNRR.so
libSPenObliquePen.so
libSPenPencil.so
libSPenPluginFW.so
libSPenScreenCodecDecoder.so
libSPenSkia.so
libSPenVIRecognition.so
libSPenVITextAll.so
libtlc_tima_atn.so
libtlc_tima_pkm.so
libtlc_tima_tui.so
libtscalendar.so
libunwind-ptrace.so
libwpa_client.so
libxml2.so